chore(fe): rename mock data file
